requirements.txt
amqp==5.0.6
billiard==3.6.4.0
cachelib==0.9.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.0.1
Flask-Caching==2.0.2
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.2
MarkupSafe==2.1.2
numpy==1.23.5
pandas==1.5.3
prompt-toolkit==3.0.38
python-dateutil==2.8.2
pytz==2023.3
redis==3.5.3
SimpleITK
numba
Werkzeug
six
vine
wcwidth
zipp
tqdm
torch
torchvision
torchaudio
torchmetrics
#py-spy
